ai-dlc/agentic_interviewer_ai/packages/entrevista-shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "entrevista-shared"
version = "0.1.0"
description = "Módulos compartidos para los lambdas de EntreVista AI"
requires-python = ">=3.12"
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "aws-xray-sdk>=2.12.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["entrevista_shared"]
